requirements.txt
flask
twilio
openai
Pillow
requests
uuid
pandas
zenrows
scrapy
inference_sdk 
Python-IO
pytesseract==0.3.10

lxml
selenium
setuptools
html5lib
pypdf
bs4
pypdf2
ai71
langchain
langchain_community
sentence-transformers
chromadb
PyMuPDF